bug endereco

+++ zghost/bus/Device.hpp
@@ -6,9 +6,7 @@
 
 class Device {
   public:
-    // Device() {}
-    // virtual ~Device() {}
-    virtual std::tuple<uint8_t, bool> read(const uint8_t& address) = 0;
+    virtual std::tuple<uint8_t, bool> read(const uint16_t& address) = 0;
     virtual bool write(const uint16_t& address, const uint8_t& value) = 0;
     virtual bool valid(const uint16_t& address) = 0;
 };
