+++ zghost/bus/Memory.hpp
@@ -9,7 +9,7 @@ class Memory : public Device {
   public:
     Memory(const uint16_t& start, const uint16_t& size);
     virtual ~Memory();
-    virtual std::tuple<uint8_t, bool> read(const uint8_t& address) override;
+    virtual std::tuple<uint8_t, bool> read(const uint16_t& address) override;
     virtual bool write(const uint16_t& address, const uint8_t& value) override;
     virtual bool valid(const uint16_t& address) override;
